Pass through test failures, and use relative paths in resource rep
First, this CL fixes a problem where the mechanism to pass through an
exception encountered during a test run gets surfaced out to the test
infrastructure. With the partial analysis change this only worked for
the default test mode.

Once I fixed that, a hidden problem surfaced: the check to make sure we
don't have absolute paths in persistence files actually fails for the
resource repository persistence files! This CL therefore also fixes the
resource repository to use the path variable mechanism to create
relative paths such that these can be shared in the remote cache. (Note
however that the plan is to ditch lint's own resource repository
implementation with the standard Studio one which is being moved up to
tools/base.)

Change-Id: Ia964ded192fd5d80f35eee85929052fab8af55c9