Add the studio-checks lint project
We have a number of lint checks intended to be run on our own codebase.
These are developed in the lint/studio-checks project, but that project
was only built by bazel. This CL also registers this project with Gradle
such that it's part of the tools/ project, and participates in find
usages, refactoring, you can run its unit tests, etc. The project does
not apply any publishing plugins; it's not meant to be distributed.

It also adds an application plugin to the lint cli project such that you
can run

    ./gradlew :base:lint-cli:install

to make a local launchable installation of lint.

Bug: None
Test: N/A
Change-Id: Id824d3daefc78fce63ea9d39f22a2ea9dc95db28