[Lint] Allow missing compileOnly dependencies in resolveDependencies()
Consider the case of libA, which has an implementation dependency on
libB, which has a compileOnly dependency on libC. Previously, when
running `./gradlew :libA:lintDebug` with checkDependencies true, lint
would throw an error from LintModelModuleProject.resolveDependencies()
because libC would be listed as a dependency in libB's lint model, but
libC's lint model wouldn't be passed to lint via --lint-model when
invoked from libA (because compileOnly dependencies are not transitive).

With this change, LintModelModuleProject.resolveDependencies() does not
throw an error for compileOnly dependencies.

This change also improves the error message by (1) replacing
project.name with project.getName() because project.name was returning
null and (2) changing the wording to make it more accurate.

Fixes: 191296077
Test: LintCompileOnlyTest
Change-Id: I5e42520db91833e0492222714f50b9aab4a831e4