adblib: Increase timeouts for tests
When running adblib tests in a loop on high end machines,
specifically, `bazel ... --runs_per_test=1000` on a Windows
machine with 32 physical cores, some tests sporadically fail
due to expiring timeouts.

One of the problematic timeout is the socket connection
timeout, which is increased from 5 seconds to 2 minutes.
This is only required on Windows where it seems the operating
system takes a much longer time to initialize socket connections
to the localhost address (127.0.0.1) when there are a lot of
sockets still "active" in the TIME_WAIT state. Windows is by
default limited to 4,000 ephemeral sockets, and sockets remain
in the TIME_WAIT for about 4 minutes after being closed. Given
that running adblib test thousand of times create thousand of
sockets, we quickly get close to the limit, and it looks like
Windows throttles socket connections when getting close to tha
limit.

The other problematic timeout is the "connect a fake device
with fake adb server" request. This is an in-process request
that goes through a sequential executor inside  fake adb server.
It is not clear why a timeout of 1 second is not enough
in some cases (i.e. we don't know why the sequential executor
does not process pending request fast enough in some case),
but increasing to 120 seconds makes stress testing 100% reliable.

Bug: 188067646
Test: this
Change-Id: Id0d6d649ea3a736366c3d8295d68bbad265a2515