Support setting lintOptions.textOutput with 'stderr' and 'stdout'
This change fixes a typo which caused nothing to be printed to stderr
when lintOptions.textOutput was set with 'stderr'.

After fixing the typo and adding the LintTextOutputTest, the test was
still failing on Windows in a way that indicated neither 'stdout' nor
'stderr' worked on Windows. It was broken on Windows  because
file.getPath() wasn't returning the original string passed to the File
constructor. I modified the File.isLintStdout() and
File.isLintStderr() methods so that they would work on Windows.

Fixes: 192090767
Test: LintTextOutputTest
Change-Id: Ida91341675599eb1cbc5248d8b58baeb4f3e0ef4